Remove redundant code execution and checks.

+++ xproc.cpp
@@ -424,12 +424,15 @@ namespace ngs::xproc {
   }
 
   bool proc_id_exists(PROCID proc_id) {
-    std::vector<PROCID> vec = proc_id_enum();
+    std::vector<PROCID> vec;
+    if (proc_id < 0) return vec;
+    vec = proc_id_enum();
     auto itr = std::find(vec.begin(), vec.end(), proc_id);
     return (itr != vec.end());
   }
 
   bool proc_id_suspend(PROCID proc_id) {
+    if (proc_id < 0) return false;
     #if !defined(_WIN32)
     return (kill(proc_id, SIGSTOP) != -1);
     #else
@@ -449,6 +452,7 @@ namespace ngs::xproc {
   }
 
   bool proc_id_resume(PROCID proc_id) {
+    if (proc_id < 0) return false;
     #if !defined(_WIN32)
     return (kill(proc_id, SIGCONT) != -1);
     #else
@@ -468,6 +472,7 @@ namespace ngs::xproc {
   }
 
   bool proc_id_kill(PROCID proc_id) {
+    if (proc_id < 0) return false;
     #if !defined(_WIN32)
     return (kill(proc_id, SIGKILL) != -1);
     #else
@@ -479,8 +484,9 @@ namespace ngs::xproc {
     #endif
   }
 
-  std::vector<PROCID> parent_proc_id_from_proc_id(PROCID proc_id) {
+  std::vector<PROCID> parent_proc_id_from_proc_id(PROCID proc_id) { 
     std::vector<PROCID> vec;
+    if (proc_id < 0) return vec;
     #if defined(_WIN32)
     HANDLE hp = CreateToolhelp32Snapshot(TH32CS_SNAPPROCESS, 0);
     if (!hp) return vec;
@@ -581,6 +587,7 @@ namespace ngs::xproc {
 
   std::vector<PROCID> proc_id_from_parent_proc_id(PROCID parent_proc_id) {
     std::vector<PROCID> vec;
+    if (proc_id < 0) return vec;
     #if defined(_WIN32)
     HANDLE hp = CreateToolhelp32Snapshot(TH32CS_SNAPPROCESS, 0);
     if (!hp) return vec;
@@ -691,6 +698,8 @@ namespace ngs::xproc {
   }
 
   std::vector<PROCID> proc_id_from_exe(std::string exe) {
+    std::vector<PROCID> vec;
+    if (exe.empty()) return vec;
     auto fnamecmp = [](std::string fname1, std::string fname2) {
       #if defined(_WIN32)
       std::size_t fp = fname2.find_last_of("\\/");
@@ -709,7 +718,6 @@ namespace ngs::xproc {
       if (abspath) return (fname1 == fname2 || fname1 == fname2.substr(0, fp));
       return (fname1 == fname2.substr(fp + 1));
     };
-    std::vector<PROCID> vec;
     std::vector<PROCID> proc_id = proc_id_enum();
     for (std::size_t i = 0; i < proc_id.size(); i++) {
       if (fnamecmp(exe, exe_from_proc_id(proc_id[i]))) {
@@ -720,11 +728,12 @@ namespace ngs::xproc {
   }
 
   std::vector<PROCID> proc_id_from_cwd(std::string cwd) {
+    std::vector<PROCID> vec;
+    if (exe.empty()) return vec;
     auto fnamecmp = [](std::string fname1, std::string fname2) {
       if (fname1.empty() || fname2.empty()) return false;
       return (fname1 == fname2);
     };
-    std::vector<PROCID> vec;
     std::vector<PROCID> proc_id = proc_id_enum();
     for (std::size_t i = 0; i < proc_id.size(); i++) {
       if (fnamecmp(cwd, cwd_from_proc_id(proc_id[i]))) {
@@ -736,6 +745,7 @@ namespace ngs::xproc {
 
   std::string exe_from_proc_id(PROCID proc_id) {
     std::string path;
+    if (proc_id < 0) return path;
     #if defined(_WIN32)
     if (proc_id == GetCurrentProcessId()) {
       wchar_t buffer[MAX_PATH];
@@ -892,6 +902,7 @@ namespace ngs::xproc {
 
   std::string cwd_from_proc_id(PROCID proc_id) {
     std::string path;
+    if (proc_id < 0) return path;
     #if defined(_WIN32)
     HANDLE proc = open_process_with_debug_privilege(proc_id);
     if (proc == nullptr) return path;
@@ -1009,6 +1020,7 @@ namespace ngs::xproc {
 
   std::vector<std::string> cmdline_from_proc_id(PROCID proc_id) {
     std::vector<std::string> vec;
+    if (proc_id < 0) return vec;
     #if defined(_WIN32)
     HANDLE proc = open_process_with_debug_privilege(proc_id);
     if (proc == nullptr) return vec;
@@ -1124,6 +1136,7 @@ namespace ngs::xproc {
 
   std::vector<std::string> environ_from_proc_id(PROCID proc_id) {
     std::vector<std::string> vec;
+    if (proc_id < 0) return vec;
     #if defined(_WIN32)
     HANDLE proc = open_process_with_debug_privilege(proc_id);
     if (proc == nullptr) return vec;
@@ -1240,6 +1253,7 @@ namespace ngs::xproc {
 
   std::string envvar_value_from_proc_id(PROCID proc_id, std::string name) {
     std::string value;
+    if (proc_id < 0 || name.empty()) return value;
     std::vector<std::string> vec = environ_from_proc_id(proc_id);
     if (!vec.empty()) {
       for (std::size_t i = 0; i < vec.size(); i++) {
@@ -1262,6 +1276,7 @@ namespace ngs::xproc {
 
   bool envvar_exists_from_proc_id(PROCID proc_id, std::string name) {
     bool exists = false;
+    if (proc_id < 0 || name.empty()) return exists;
     std::vector<std::string> vec = environ_from_proc_id(proc_id);
     if (!vec.empty()) {
       for (std::size_t i = 0; i < vec.size(); i++) {
